Created simple landing page with log in button; Made first fetch from Spotify Web API
When there is not already user information known, log in button will show. This button will redirect to the backend app and allows the user to log in with their spotify details, and then be redirected back to the main display page with an access token. First fetch from Spotifys API used to display current users display name in the title