Added functionality to drop down button on Item section - Component name change FavouritesItem -> DataItem
FavouritesItem component name changed to DataItem (keeps it more general since its not just in the favourites section now)

Functionality added to the drop down button in DataItem component. The button works as a toggle. Condition returns the hidden section if this.state.toggled is true, otherwise it returns null, therefore not rendering anything