Fixed spacing in CurrentlyPlayed and its height inside the frame; Changed style of buttons; Replaced the rotated skipButton svg with new backButton SVG (caused problems with other stlying properties); Went from min font size 16px to 12px for general content text;
CurrentlyPlayed progress bar now sits at the bottom of the frame, height of the section is now fixed to fit the frame.
Gave the MediaControls and song and artist text more space within the CurrentlyPlaying section.
Changed the styling of the MediaControl buttons to black background, with orange fill and border