Added play pause track functionality to the MediaControls component
The button and playback state has some funky interactions with the API that I will iron out but is not a priority right now.

For now. Playback must be started from another source.
Once a current track is recognised to be playing by the app, playback can be toggled intuitively with the buttons.

Buttons are separate for now but will be converted to toggle like behaviour and will only take up one button slot.